requirements.txt
cycler==0.11.0
fonttools==4.28.5
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
Pillow==9.0.0
pykeops==1.5
pyparsing==3.0.6
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
cycler==0.11.0
decorator==5.1.1
fonttools==4.28.5
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
Pillow==9.0.0
pykeops==1.5
pyparsing==3.0.6
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
validators==0.18.2
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
decorator==5.1.1
fonttools==4.28.5
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
Pillow==9.0.0
pykeops==1.5
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
urllib3==1.26.8
validators==0.18.2
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
decorator==5.1.1
fonttools==4.28.5
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.0
packaging==21.3
Pillow==9.0.0
pykeops==1.5
pyparsing==3.0.6
python-dateutil==2.8.2
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
urllib3==1.26.8
validators==0.18.2
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
decorator==5.1.1
fastapi==0.68.2
fonttools==4.28.5
grpcio==1.43.0
h11==0.13.0
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
mlserver==0.5.3
mlserver-sklearn==0.5.3
numpy==1.22.0
packaging==21.3
pandas==1.3.5
Pillow==9.0.0
protobuf==3.19.3
pydantic==1.9.0
pykeops==1.5
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
starlette==0.14.2
threadpoolctl==3.0.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0
validators==0.18.2
